fix #291 group order
